requirements.txt
Django==4.2.9
django-environ==0.11.2
djangorestframework==3.14.0
pandas==2.2.0
psycopg2==2.9.9
requests==2.31.0
SQLAlchemy==2.0.25
